requirements.txt
azure-storage-blob
azure-identity
azure-mgmt-resource
azure-mgmt-storage
flask
requests
Werkzeug